+++ packages/snaps-rpc-methods/src/restricted/manageState.test.ts
@@ -65,6 +65,13 @@ describe('snap_manageState', () => {
   });
 
   describe('getManageStateImplementation', () => {
+    if (!('CryptoKey' in globalThis)) {
+      // We can remove this once we drop Node 18
+      Object.defineProperty(globalThis, 'CryptoKey', {
+        value: webcrypto.CryptoKey,
+      });
+    }
+
     it('gets snap state', async () => {
       const mockSnapState = {
         some: {