+++ packages/snaps-utils/src/checksum.test.ts
@@ -2,7 +2,7 @@ import * as nobleHashes from '@noble/hashes/sha256';
 import { base64 } from '@scure/base';
 import { webcrypto } from 'crypto';
 
-import { checksum, checksumFiles, getChecksumBytes } from './checksum';
+import { checksum, checksumFiles } from './checksum';
 import { VirtualFile } from './index.browser';
 
 const EMPTY_SHA256 = '47DEQpj8HBSa+/TImW+5JCeuQeRkm5NMpJWZG3hSuFU=';
@@ -137,24 +137,5 @@ describe('checkumFiles', () => {
     const hash2 = base64.encode(await checksumFiles(files2));
 
     expect(hash1).toBe(hash2);
-  });
-});
-
-describe('getChecksumBytes', () => {
-  it('handles Uint8Array', () => {
-    expect(getChecksumBytes(FOO_BAR_UINT8)).toStrictEqual(FOO_BAR_UINT8);
-  });
-
-  it('handles strings', () => {
-    expect(getChecksumBytes(FOO_BAR_STR)).toStrictEqual(FOO_BAR_UINT8);
-  });
-
-  it('handles virtual files', () => {
-    expect(getChecksumBytes(new VirtualFile(FOO_BAR_UINT8))).toStrictEqual(
-      FOO_BAR_UINT8,
-    );
-    expect(getChecksumBytes(new VirtualFile(FOO_BAR_STR))).toStrictEqual(
-      FOO_BAR_UINT8,
-    );
   });
 });